Custom TableSortLabel with brighter arrow color
